requirements.txt
asgiref==3.5.2
Django~=4.1.0
sqlparse==0.4.2
Pillow==9.2.0
flower==1.2.0
celery==5.2.7
stripe==4.0.2
WeasyPrint==56.1
redis==4.3.4
django-rosetta==0.9.8
django-parler==2.3
django-localflavor==3.1
django-environ==0.9.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
black==22.6.0
daphne==3.0.2
django-redisboard==8.3.0
requests==2.28.1
whitenoise==6.2.0
rosetta==0.3
